requirements/base.txt
Django==5.1.3
djangorestframework==3.15.2
django-modeltranslation-0.19.11
Pillow-11.0.0